+++ src/core_modules/capture-core/components/DataEntries/Enrollment/EnrollmentDataEntry.component.js
@@ -44,6 +44,7 @@ import {
 } from '../../DataEntryDhis2Helpers';
 import { convertDateObjectToDateFormatString } from '../../../../capture-core/utils/converters/date';
 import { systemSettingsStore } from '../../../metaDataMemoryStores';
+import type { RelatedStageRefPayload } from '../../WidgetRelatedStages';
 
 const overrideMessagePropNames = {
     errorMessage: 'validationError',
@@ -247,7 +248,7 @@ const getGeometrySettings = () => ({
                 dialogLabel: i18n.t('Area'),
                 required: false,
                 orientation: getOrientation(props.formHorizontal),
-                orgUnit: props.orgUnit,
+                orgUnitId: props.orgUnit?.id,
             });
         }
 
@@ -258,7 +259,7 @@ const getGeometrySettings = () => ({
             required: false,
             orientation: getOrientation(props.formHorizontal),
             shrinkDisabled: props.formHorizontal,
-            orgUnit: props.orgUnit,
+            orgUnitId: props.orgUnit?.id,
         });
     },
     getPropName: () => 'geometry',
@@ -339,10 +340,12 @@ type FinalTeiDataEntryProps = {
     programId: string,
     id: string,
     orgUnitId: string,
+    orgUnit: OrgUnit,
     onUpdateDataEntryField: Function,
     onUpdateFormFieldAsync: Function,
     onUpdateFormField: Function,
     firstStageMetaData?: ?{ stage: ProgramStage },
+    relatedStageRef?: { current: ?RelatedStageRefPayload },
     formFoundation: RenderFoundation,
 };
 // final step before the generic dataEntry is inserted
@@ -467,6 +470,7 @@ export class EnrollmentDataEntryComponent extends React.Component<PreEnrollmentD
             onUpdateDataEntryField,
             onStartAsyncUpdateField,
             onGetUnsavedAttributeValues,
+            orgUnit,
             ...passOnProps
         } = this.props;
 
@@ -477,6 +481,8 @@ export class EnrollmentDataEntryComponent extends React.Component<PreEnrollmentD
                 onUpdateFormField={this.handleUpdateField}
                 onUpdateDataEntryField={this.handleUpdateDataEntryField}
                 onUpdateFormFieldAsync={this.handleStartAsyncUpdateField}
+                orgUnit={orgUnit}
+                orgUnitId={orgUnit?.id}
                 {...passOnProps}
             />
         );