+++ src/core_modules/capture-core/components/DataEntries/Enrollment/EnrollmentWithFirstStageDataEntry/EnrollmentWithFirstStageDataEntry.component.js
@@ -129,7 +129,7 @@ const getStageGeometrySettings = () => ({
                 dialogLabel: i18n.t('Area'),
                 required: false,
                 orientation: getOrientation(props.formHorizontal),
-                orgUnit: props.orgUnit,
+                orgUnitId: props.orgUnit?.id,
             });
         }
 
@@ -140,7 +140,7 @@ const getStageGeometrySettings = () => ({
             required: false,
             orientation: getOrientation(props.formHorizontal),
             shrinkDisabled: props.formHorizontal,
-            orgUnit: props.orgUnit,
+            orgUnitId: props.orgUnit?.id,
         });
     },
     getPropName: () => stageMainDataIds.GEOMETRY,