+++ src/core_modules/capture-core/components/DataEntries/SingleEventRegistrationEntry/DataEntryWrapper/DataEntry/epics/addNoteForNewSingleEvent.epics.js
@@ -2,10 +2,7 @@
 import uuid from 'd2-utilizr/lib/uuid';
 import { ofType } from 'redux-observable';
 import { switchMap } from 'rxjs/operators';
-
 import moment from 'moment';
-import { convertValue as convertListValue } from '../../../../../../converters/clientToList';
-import { dataElementTypes } from '../../../../../../metaData';
 import {
     actionTypes as newEventDataEntryActionTypes,
 } from '../actions/dataEntry.actions';
@@ -28,7 +25,6 @@ export const addNoteForNewSingleEventEpic = (action$: InputObservable, store: Re
             }).then((user) => {
                 const { userName, firstName, surname } = user;
                 const clientId = uuid();
-                const storedAt = moment().toISOString();
                 const note = {
                     value: payload.note,
                     createdBy: {