+++ src/core_modules/capture-core/components/WidgetRelatedStages/WidgetRelatedStages.types.js
@@ -1,5 +1,5 @@
 // @flow
-import { actions as LinkModes } from './constants';
+import { relatedStageActions } from './index';
 import type { Constraint } from './RelatedStagesActions/RelatedStagesActions.types';
 
 export type RelationshipType = {|
@@ -17,12 +17,18 @@ export type RelationshipType = {|
 
 export type Props = {|
     programId: string,
-    enrollmentId: string,
+    enrollmentId?: string,
     programStageId: string,
-    currentStageLabel: string,
+    actionsOptions?: {
+        [key: $Keys<typeof relatedStageActions>]: {
+            hidden?: boolean,
+            disabled?: boolean,
+            disabledMessage?: string
+        },
+    },
 |}
 export type RelatedStageDataValueStates = {|
-    linkMode: ?$Keys<typeof LinkModes>,
+    linkMode: ?$Keys<typeof relatedStageActions>,
     scheduledAt: string,
     scheduledAtFormatError: ?{error: ?string, errorCode: ?string},
     orgUnit: ?{
@@ -32,3 +38,27 @@ export type RelatedStageDataValueStates = {|
     },
     linkedEventId: ?string,
 |}
+
+export type RelatedStageRelationshipType = {|
+    id: string,
+    fromConstraint: {|
+        programStage: {
+            id: string,
+        },
+    |},
+    toConstraint: {
+        programStage: {
+            id: string,
+        },
+    }
+|}
+
+export type RelatedStageRefPayload = {
+    getLinkedStageValues: () => {
+        selectedRelationshipType: RelatedStageRelationshipType,
+        relatedStageDataValues: RelatedStageDataValueStates,
+        linkMode: ?$Keys<typeof relatedStageActions>,
+    },
+    eventHasLinkableStageRelationship: () => boolean,
+    formIsValidOnSave: () => boolean,
+};