+++ src/core_modules/capture-core/components/WidgetStagesAndEvents/types/common.types.js
@@ -6,6 +6,7 @@ import { dataElementTypes, Option } from '../../../metaData';
 type StageOptions = {
     [code: string]: string;
 }
+
 export type StageDataElement = {
     id: string,
     name: string,
@@ -14,6 +15,15 @@ export type StageDataElement = {
     options?: StageOptions,
     optionSet?: { options: Array<Option> },
 }
+
+export type StageDataElementClient = {
+    id: string,
+    name: string,
+    formName: string,
+    type: $Keys<typeof dataElementTypes>,
+    options?: Array<{ value: any, text: any }>,
+    optionSet?: { options: Array<any> },
+};
 
 export type Stage = {
     id: string,