+++ src/core_modules/capture-core/utils/converters/date/index.js
@@ -3,5 +3,3 @@ export { convertDateObjectToDateFormatString } from './dateObjectToDateFormatStr
 export { convertMomentToDateFormatString } from './momentToDateFormatString';
 export { convertIsoToLocalCalendar } from './convertIsoToLocalCalendar';
 export { convertLocalToIsoCalendar } from './convertLocalToIsoCalendar';
-export { convertStringToTemporal } from './convertStringToTemporal';
-export { convertTemporalToString } from './convertTemporalToString';