+++ src/core_modules/capture-ui/CoordinateField/CoordinateField.component.js
@@ -20,7 +20,8 @@ type Coordinate = {
 
 type Props = {
   onBlur: (value: any) => void,
-  onOpenMap: (hasValue: boolean) => void,
+  onOpenMap?: (hasValue: boolean) => void,
+  onCloseMap?: () => void,
   orientation: $Values<typeof orientations>,
   center?: ?Array<number>,
   onChange?: ?(value: any) => void,
@@ -89,11 +90,12 @@ export class CoordinateField extends React.Component<Props, State> {
     }
 
     closeMap = () => {
+        this.props.onCloseMap?.();
         this.setState({ showMap: false });
     }
 
     openMap = () => {
-        this.props.onOpenMap(Boolean(this.props.value));
+        this.props.onOpenMap?.(Boolean(this.props.value));
         this.setState({ showMap: true, position: this.getPosition() });
     }
 
@@ -169,7 +171,10 @@ export class CoordinateField extends React.Component<Props, State> {
     }
 
     renderMap = () => {
-        const { position, zoom } = this.state;
+        const { position, zoom, showMap } = this.state;
+        if (!showMap || (!position && !this.props.center)) {
+            return null;
+        }
         const center = position || this.props.center;
         return (
             <div className={defaultClasses.mapContainer}>